fixes #41 Test coverage for entity model  -  Moves all routing methods to test_routing
